Allow gnudoc to rebuild info files
gendocs.sh from gnulib absolutely insists upon rebuilding the .info
files in the source directory. Unfortunately, distcheck demands that
that directory remain unmodified during that check. So, put the
invocation into a script and carefully enable and disable the
doc directory for the duration of that script. Lovely.

remove debugging stuff and generated files

Files changed:
* bootstrap: update
* bootstrap.conf: update
* bootstrap.std: update
* doc/Makefile.am: rework to use rewriten script
* doc/mk.sh: renamed to mk-texi.sh and edited
* doc/complexity-tpl.texi: renamed to example-text.tpl
* doc/complexity.texi: new file extracted from complexity-tpl.txt
* doc/example-texi.tpl: edited from complexity-tpl.texi